+++ wazo_agentd/http.py
@@ -289,12 +289,12 @@ class _UnpauseAgentByNumber(_BaseResource):
         return '', 204
 
 
-class _ResumeUserAgent(_BaseResource):
-    @required_acl('agentd.users.me.agents.resume.create')
+class _UnpauseUserAgent(_BaseResource):
+    @required_acl('agentd.users.me.agents.unpause.create')
     def post(self):
         tenant_uuids = self._build_tenant_list({'recurse': True})
         user_uuid = token.user_uuid
-        self.service_proxy.resume_user_agent(user_uuid, tenant_uuids=tenant_uuids)
+        self.service_proxy.unpause_user_agent(user_uuid, tenant_uuids=tenant_uuids)
         return '', 204
 
 
@@ -313,7 +313,7 @@ class HTTPInterface:
         (_LogoffAgentByNumber, '/agents/by-number/<agent_number>/logoff'),
         (_LogoffUserAgent, '/users/me/agents/logoff'),
         (_PauseUserAgent, '/users/me/agents/pause'),
-        (_ResumeUserAgent, '/users/me/agents/resume'),
+        (_UnpauseUserAgent, '/users/me/agents/unpause'),
         (_AddAgentToQueue, '/agents/by-id/<int:agent_id>/add'),
         (_RemoveAgentFromQueue, '/agents/by-id/<int:agent_id>/remove'),
         (_PauseAgentByNumber, '/agents/by-number/<agent_number>/pause'),